Refactor RawProducer and RawConsumer classes
Remove duplication code between the Raw* classes and moved to
AbstractEntity class.

Fix a rare condition where the producer still receive confirmation but
the producer was removed from the internal list due of cancellation.

+++ RabbitMQ.Stream.Client/Client.cs
@@ -331,22 +331,28 @@ namespace RabbitMQ.Stream.Client
             return (publisherId, response);
         }
 
-        public async Task<DeletePublisherResponse> DeletePublisher(byte publisherId)
+        public async Task<DeletePublisherResponse> DeletePublisher(byte publisherId,
+            bool ignoreIfAlreadyRemoved = false)
         {
             await _poolSemaphore.WaitAsync().ConfigureAwait(false);
             try
             {
-                var result =
-                    await Request<DeletePublisherRequest, DeletePublisherResponse>(corr =>
-                        new DeletePublisherRequest(corr, publisherId)).ConfigureAwait(false);
+                if (!ignoreIfAlreadyRemoved)
+                {
+                    var result =
+                        await Request<DeletePublisherRequest, DeletePublisherResponse>(corr =>
+                            new DeletePublisherRequest(corr, publisherId)).ConfigureAwait(false);
 
-                return result;
+                    return result;
+                }
             }
             finally
             {
                 publishers.Remove(publisherId);
                 _poolSemaphore.Release();
             }
+
+            return new DeletePublisherResponse();
         }
 
         public async Task<(byte, SubscribeResponse)> Subscribe(string stream, IOffsetType offsetType,
@@ -492,12 +498,25 @@ namespace RabbitMQ.Stream.Client
                 case PublishConfirm.Key:
                     PublishConfirm.Read(frame, out var confirm);
                     confirmFrames += 1;
-                    var (confirmCallback, _) = publishers[confirm.PublisherId];
-                    confirmCallback(confirm.PublishingIds);
-                    if (MemoryMarshal.TryGetArray(confirm.PublishingIds, out var confirmSegment))
+                    if (publishers.TryGetValue(confirm.PublisherId, out var publisherConf))
                     {
-                        if (confirmSegment.Array != null)
-                            ArrayPool<ulong>.Shared.Return(confirmSegment.Array);
+                        var (confirmCallback, _) = publisherConf;
+                        confirmCallback(confirm.PublishingIds);
+                        if (MemoryMarshal.TryGetArray(confirm.PublishingIds, out var confirmSegment))
+                        {
+                            if (confirmSegment.Array != null)
+                                ArrayPool<ulong>.Shared.Return(confirmSegment.Array);
+                        }
+                    }
+                    else
+                    {
+                        // the producer is not found, this can happen when the producer is closing
+                        // and there are still confirmation on the wire 
+                        // we can ignore the error since the producer does not exists anymore
+                        _logger?.LogDebug(
+                            "Could not find stream producer {ID} or producer is closing." +
+                            "A possible cause it that the producer was closed and the are still confirmation on the wire. ",
+                            confirm.PublisherId);
                     }
 
                     break;
@@ -522,8 +541,22 @@ namespace RabbitMQ.Stream.Client
                     break;
                 case PublishError.Key:
                     PublishError.Read(frame, out var error);
-                    var (_, errorCallback) = publishers[error.PublisherId];
-                    errorCallback(error.PublishingErrors);
+                    if (publishers.TryGetValue(error.PublisherId, out var publisher))
+                    {
+                        var (_, errorCallback) = publisher;
+                        errorCallback(error.PublishingErrors);
+                    }
+                    else
+                    {
+                        // the producer is not found, this can happen when the producer is closing
+                        // and there are still confirmation on the wire 
+                        // we can ignore the error since the producer does not exists anymore
+                        _logger?.LogDebug(
+                            "Could not find stream producer {ID} or producer is closing." +
+                            "A possible cause it that the producer was closed and the are still confirmation on the wire. ",
+                            error.PublisherId);
+                    }
+
                     break;
                 case MetaDataUpdate.Key:
                     MetaDataUpdate.Read(frame, out var metaDataUpdate);
